Updated Babel Presets
I added .browserslistrc to specify the criteria Babel should support
when transpiling JS during the build process. I removed the target of
IE9 that was configured and replaced it with criteria that eliminates
all dead browsers, meaning we only have to support IE 11 now,
but that may be considered a dead browser now.

Signed-off-by: Scott Simontis <southernsadboi@protonmail.ch>